Replace cache and add http client param
The user can now pass their own httpx client if necessary.
This avoids the trouble of passing around different configs,
and additionally, opens the door for making the underlying
http client act more as an interface. For example, the user
may want to use "requests", "aiohttp", or some other http lib.
The caching option was too "hacky" and so removed from the package methods.