Remove caching mechanism for dns client
At this time, the ability to cache dns loads from iana
or even a dns client instance seems like premature optimization.
Moreover, the assumptions that need to be made for "what" httpx kwargs
look like or the data types they have is too "gray".
Other avenues such as the singleton pattern may be explored in the
near future.